+++ common/cosmwasm-smart-contracts/coconut-bandwidth-contract/src/msg.rs
@@ -5,7 +5,7 @@ use cosmwasm_std::Coin;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use crate::deposit::DepositData;
+use crate::{deposit::DepositData, spend_credential::SpendCredentialData};
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 pub struct InstantiateMsg {
@@ -17,12 +17,21 @@ pub struct InstantiateMsg {
 #[serde(rename_all = "snake_case")]
 pub enum ExecuteMsg {
     DepositFunds { data: DepositData },
+    SpendCredential { data: SpendCredentialData },
     ReleaseFunds { funds: Coin },
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
-pub enum QueryMsg {}
+pub enum QueryMsg {
+    GetSpentCredential {
+        blinded_serial_number: String,
+    },
+    GetAllSpentCredentials {
+        limit: Option<u32>,
+        start_after: Option<String>,
+    },
+}
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "snake_case")]