+++ validator-api/src/coconut/error.rs
@@ -63,22 +63,17 @@ pub enum CoconutError {
     #[error("Error in coconut interface - {0}")]
     CoconutInterfaceError(#[from] coconut_interface::error::CoconutInterfaceError),
 
-    #[error("Could not create proposal for spending credential")]
-    CreateProposalError,
-
     #[error("Storage error - {0}")]
     StorageError(#[from] ValidatorApiStorageError),
 
     #[error("Credentials error - {0}")]
     CredentialsError(#[from] credentials::error::Error),
 
-    #[error(
-        "Incorrect credential proposal description. Expected blinded serial number in base 58"
-    )]
-    IncorrectProposal,
+    #[error("Incorrect credential proposal description: {reason}")]
+    IncorrectProposal { reason: String },
 
-    #[error("Internal error: {0}")]
-    InternalError(String),
+    #[error("Invalid status of credential: {status}")]
+    InvalidCredentialStatus { status: String },
 }
 
 impl<'r, 'o: 'r> Responder<'r, 'o> for CoconutError {